// packages/epub-engine/src/types.ts
export interface LayoutSettings {
  layout?: string
  spread?: string
  minSpreadWidth?: number
  evenSpreads?: boolean
  flow?: string
  direction?: string
}
// ...
export type RequestFunction = (
  url: string,
  type?: string,
  withCredentials?: boolean,
  headers?: Record<string, string>,
  signal?: AbortSignal,
) => Promise<unknown>